Implement file system IPC
The file system layout is like this:

requests/  - request files. The filename format is <id>.json
responses/ - response files. The filename matches the request filename.

Stale requests and response files (older than 1 minute) are removed
regularly. This assumes that SM doesn't wait for responses for more than
a minute.

The fsipc module can be used like this:

	ipc, err := fsipc.New("data/folder")
	if err != nil {
		log.Fatal(err)
	}
	defer ipc.Close()

	for request := range ipc.Requests {
		switch req := request.(type) {
		case fsipc.PingRequest:
			// handle request
			...
			ipc.WriteResponse(req.Id, response)
		case fsipc.SubmitScoreRequest:
			// handle request
			...
			ipc.WriteResponse(req.Id, response)
		default:
			log.Fatal("unknown request")
		}
	}

The response can be any type that is supported by json.Marshal().